Refactored the entire project into three separate modules We now have a root module that works as a parent for the two other modules. The second module contains the MCP PKI library itself, and the third one contains the MCP PKI CLI which itself has a dependency on the MCP PKI library. This was done to make it easier to separate dependencies that are needed for the PKI and dependencies that are only needed for the CLI part, such as commons-cli and slf4j-simple.
