+++ mcp-pki-cli/src/main/java/net/maritimeconnectivity/pki/cli/Main.java
@@ -1,21 +1,27 @@
 /*
  * Copyright 2017 Danish Maritime Authority.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-package net.maritimeconnectivity.pki;
+package net.maritimeconnectivity.pki.cli;
 
 import lombok.extern.slf4j.Slf4j;
+import net.maritimeconnectivity.pki.CAHandler;
+import net.maritimeconnectivity.pki.CertificateBuilder;
+import net.maritimeconnectivity.pki.CertificateHandler;
+import net.maritimeconnectivity.pki.KeystoreHandler;
+import net.maritimeconnectivity.pki.PKIConfiguration;
+import net.maritimeconnectivity.pki.PKIIdentity;
 import net.maritimeconnectivity.pki.pkcs11.P11PKIConfiguration;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
