+++ mcp-pki/src/main/java/net/maritimeconnectivity/pki/PKIConstants.java
