+++ mcp-pki/src/main/java/net/maritimeconnectivity/pki/RevocationInfo.java
