+++ mcp-pki/src/main/java/net/maritimeconnectivity/pki/exception/PKIRuntimeException.java
