+++ mcp-pki/src/main/java/net/maritimeconnectivity/pki/ocsp/CertStatus.java
