+++ mcp-pki/src/main/resources/META-INF/LICENSE
@@ -1,0 +1,1 @@
+../../../../LICENSE
