+++ mcp-pki/src/test/resources/Certificate_Myboat.pem
