+++ mcp-pki/src/test/resources/ca/mc-bimco-cert.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDrTCCAzOgAwIBAgIUTVGYYyIYnxvZTB39eY3FvvWhhn4wCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM4WhcNMjQxMjMxMjMw
+MDAwWjCBxjEkMCIGCgmSJomT8ixkAQEMFHVybjptcm46bWNsOmNhOmJpbWNvMQsw
+CQYDVQQGEwJESzEQMA4GA1UECAwHRGVubWFyazETMBEGA1UEBwwKQ29wZW5oYWdl
+bjEYMBYGA1UECgwPQklNQ08gVW5pdCBUZXN0MRgwFgYDVQQLDA9CSU1DTyBVbml0
+IFRlc3QxFzAVBgNVBAMMDkJJTUNPIEV4dHJhTmV0MR0wGwYJKoZIhvcNAQkBFg5p
+bmZvQGJpbWNvLm9yZzB2MBAGByqGSM49AgEGBSuBBAAiA2IABL3fFxuUkcwe+AsK
+b71WN7eyw/cZvf5atov5d9S7JyyAI6Y59SxpO5njrvYm+z4cu1GoYv5EoJr4XLZT
+AmrXCi0lueSyHxLGYvgjXBuhyzNP0ssdiY5kUjk8ju95Uvlvb6OB0zCB0DAPBgNV
+HRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIB5jAfBgNVHSMEGDAWgBTys67EaTpH
+uOixcyLdCrz1vi5v3zAdBgNVHQ4EFgQUe6MLHT5hNHLdRiJNO1QO6vYvBTgwbQYD
+VR0fBGYwZDBioGCgXoZcaHR0cHM6Ly9zdGFnaW5nLWFwaS5tYXJpdGltZWNsb3Vk
+Lm5ldC94NTA5L2FwaS9jZXJ0aWZpY2F0ZXMvY3JsL3Vybjptcm46bWNsOmNhOm1h
+cml0aW1lY2xvdWQwCgYIKoZIzj0EAwIDaAAwZQIxAIPQS1VjUjoonY0UcENZK580
+upez9yMHetG/FRxToKd388SZ2mokAcMGcP8Y7boERgIwYcuirIbF1ROZwYRqkDp0
+/pLXJ/e6f6GnUwPABxWYz3tX2ZLQOwVZwI1R7YruQGIH
+-----END CERTIFICATE-----
