+++ mcp-pki/src/test/resources/ca/mc-ca-chain.pem
@@ -1,0 +1,90 @@
+-----BEGIN CERTIFICATE-----
+MIIDszCCAzigAwIBAgIUEPYUxwGeLgowuVUYdiX9n7SxR68wCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM5WhcNMjQxMjMxMjMw
+MDAwWjCByzEjMCEGCgmSJomT8ixkAQEME3Vybjptcm46bWNsOmNhOmlhbGExCzAJ
+BgNVBAYTAkZSMQ4wDAYDVQQIDAVQYXJpczEbMBkGA1UEBwwSU3QgR2VybWFpbiBl
+biBMYXllMRcwFQYDVQQKDA5JQUxBIFVuaXQgVGVzdDEXMBUGA1UECwwOSUFMQSBV
+bml0IFRlc3QxFTATBgNVBAMMDElBTEEgTWVtYmVyczEhMB8GCSqGSIb3DQEJARYS
+aW5mb0BpYWxhLWFpc20ub3JnMHYwEAYHKoZIzj0CAQYFK4EEACIDYgAEtzVUt1Yu
+EnK3wYGtKdVptHqhPHxVwI9/+qk+6HabBywrbSf9+3v8NC4AXIOFqraanCbgY46j
+W8qB7+UCpbE3BawRn5ShDlsUlcdX4UOLIHCNO1Nq/7Nj90wMlbSCOXgro4HTMIHQ
+MA8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgHmMB8GA1UdIwQYMBaAFPKz
+rsRpOke46LFzIt0KvPW+Lm/fMB0GA1UdDgQWBBSCytuLNaEyJFo4oYVnhk7xG7ON
+4jBtBgNVHR8EZjBkMGKgYKBehlxodHRwczovL3N0YWdpbmctYXBpLm1hcml0aW1l
+Y2xvdWQubmV0L3g1MDkvYXBpL2NlcnRpZmljYXRlcy9jcmwvdXJuOm1ybjptY2w6
+Y2E6bWFyaXRpbWVjbG91ZDAKBggqhkjOPQQDAgNpADBmAjEAvngPvblrjcSwp6nF
+mkFhDHdCQOCqaOqecqT/zUylDVmqSp/hkchE29RMI4cXwZQxAjEAuH52vQonHhl5
+sctrq85U3e9RmTDEau5TOgmmsliGx7JJglSycsE1m0mLQh1t4Tvr
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIDrTCCAzOgAwIBAgIUTVGYYyIYnxvZTB39eY3FvvWhhn4wCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM4WhcNMjQxMjMxMjMw
+MDAwWjCBxjEkMCIGCgmSJomT8ixkAQEMFHVybjptcm46bWNsOmNhOmJpbWNvMQsw
+CQYDVQQGEwJESzEQMA4GA1UECAwHRGVubWFyazETMBEGA1UEBwwKQ29wZW5oYWdl
+bjEYMBYGA1UECgwPQklNQ08gVW5pdCBUZXN0MRgwFgYDVQQLDA9CSU1DTyBVbml0
+IFRlc3QxFzAVBgNVBAMMDkJJTUNPIEV4dHJhTmV0MR0wGwYJKoZIhvcNAQkBFg5p
+bmZvQGJpbWNvLm9yZzB2MBAGByqGSM49AgEGBSuBBAAiA2IABL3fFxuUkcwe+AsK
+b71WN7eyw/cZvf5atov5d9S7JyyAI6Y59SxpO5njrvYm+z4cu1GoYv5EoJr4XLZT
+AmrXCi0lueSyHxLGYvgjXBuhyzNP0ssdiY5kUjk8ju95Uvlvb6OB0zCB0DAPBgNV
+HRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIB5jAfBgNVHSMEGDAWgBTys67EaTpH
+uOixcyLdCrz1vi5v3zAdBgNVHQ4EFgQUe6MLHT5hNHLdRiJNO1QO6vYvBTgwbQYD
+VR0fBGYwZDBioGCgXoZcaHR0cHM6Ly9zdGFnaW5nLWFwaS5tYXJpdGltZWNsb3Vk
+Lm5ldC94NTA5L2FwaS9jZXJ0aWZpY2F0ZXMvY3JsL3Vybjptcm46bWNsOmNhOm1h
+cml0aW1lY2xvdWQwCgYIKoZIzj0EAwIDaAAwZQIxAIPQS1VjUjoonY0UcENZK580
+upez9yMHetG/FRxToKd388SZ2mokAcMGcP8Y7boERgIwYcuirIbF1ROZwYRqkDp0
+/pLXJ/e6f6GnUwPABxWYz3tX2ZLQOwVZwI1R7YruQGIH
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIID7jCCA3WgAwIBAgIUf6H773xtaSLjygXofBl+ZTtgVkowCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM4WhcNMjQxMjMxMjMw
+MDAwWjCCAQcxMjAwBgoJkiaJk/IsZAEBDCJ1cm46bXJuOm1jbDpjYTptYXJpdGlt
+ZWNsb3VkLWlkcmVnMQswCQYDVQQGEwJESzEQMA4GA1UECAwHRGVubWFyazETMBEG
+A1UEBwwKQ29wZW5oYWdlbjEgMB4GA1UECgwXTWFyaXRpbWVDbG91ZCBVbml0IFRl
+c3QxIDAeBgNVBAsMF01hcml0aW1lQ2xvdWQgVW5pdCBUZXN0MTIwMAYDVQQDDClN
+YXJpdGltZUNsb3VkIFVuaXQgVGVzdCBJZGVudGl0eSBSZWdpc3RyeTElMCMGCSqG
+SIb3DQEJARYWaW5mb0BtYXJpdGltZWNsb3VkLm5ldDB2MBAGByqGSM49AgEGBSuB
+BAAiA2IABNfcAsujrQ6GUDnMjKy0Uo0RR00eRn+2vVzb6687+Ozk0WQSlaitKU6G
+3RlPKniWUU0JBsXCk1QY4oPPy6TmnY2z8tQmKra52dWVwPSYsrnvHyKnjOEnKMKu
+u4GXJAdasqOB0zCB0DAPBgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIB5jAf
+BgNVHSMEGDAWgBTys67EaTpHuOixcyLdCrz1vi5v3zAdBgNVHQ4EFgQUhYEdnD6m
+UQHPbIA3WrBepBvwJYMwbQYDVR0fBGYwZDBioGCgXoZcaHR0cHM6Ly9zdGFnaW5n
+LWFwaS5tYXJpdGltZWNsb3VkLm5ldC94NTA5L2FwaS9jZXJ0aWZpY2F0ZXMvY3Js
+L3Vybjptcm46bWNsOmNhOm1hcml0aW1lY2xvdWQwCgYIKoZIzj0EAwIDZwAwZAIw
+JZdNXFvE9HHo3lJqFk779HXXzRKlPXXtMMS5BP511c11y8c316kGohhEdmL4WLht
+AjAngKHJPREwbqgrikHsdjm75XopIkctYdKm4qLvuw+NoB22e9YWGRCwrVa8PQhl
+oZ8=
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIDujCCAz+gAwIBAgIUDhlcvo7GkIJ5Z9aJX2f9PuSchgowCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM3WhcNMjQxMjMxMjMw
+MDAwWjCB0jELMAkGA1UEBhMCREsxEDAOBgNVBAgMB0Rlbm1hcmsxEzARBgNVBAcM
+CkNvcGVuaGFnZW4xIDAeBgNVBAoMF01hcml0aW1lQ2xvdWQgVW5pdCBUZXN0MSAw
+HgYDVQQLDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDExMC8GA1UEAwwoTWFyaXRp
+bWVDbG91ZCBVbml0IFRlc3QgUm9vdCBDZXJ0aWZpY2F0ZTElMCMGCSqGSIb3DQEJ
+ARYWaW5mb0BtYXJpdGltZWNsb3VkLm5ldDB2MBAGByqGSM49AgEGBSuBBAAiA2IA
+BIBZiJcroN0p+tx+KjQC47lyPIAul9jdj4mT5sFbskcIWlazhGTpC42Fctsikox8
+jIxjfGQCLIr15/VDymMLM8PT2gYXjgWAWCsLgvFHFvwSTlvpd4vbewY/nBJgOMz1
+haOB0zCB0DAPBgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIB5jAfBgNVHSME
+GDAWgBTys67EaTpHuOixcyLdCrz1vi5v3zAdBgNVHQ4EFgQU8rOuxGk6R7josXMi
+3Qq89b4ub98wbQYDVR0fBGYwZDBioGCgXoZcaHR0cHM6Ly9zdGFnaW5nLWFwaS5t
+YXJpdGltZWNsb3VkLm5ldC94NTA5L2FwaS9jZXJ0aWZpY2F0ZXMvY3JsL3Vybjpt
+cm46bWNsOmNhOm1hcml0aW1lY2xvdWQwCgYIKoZIzj0EAwIDaQAwZgIxAOWDOybk
+7mCQEcXae3DFfFhTZ9KtUg9WwoneRrcvyNYGrVSwolumWJbs6EpATKxyZQIxALs4
+WqZr+pBKHQpJuCvdJ/fDWSrHIb0kUS0B27LNNVbTODzE35mKGptcGFkbVIHwfA==
+-----END CERTIFICATE-----
