+++ mcp-pki/src/test/resources/ca/mc-idreg-cert.pem
@@ -1,0 +1,24 @@
+-----BEGIN CERTIFICATE-----
+MIID7jCCA3WgAwIBAgIUf6H773xtaSLjygXofBl+ZTtgVkowCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM4WhcNMjQxMjMxMjMw
+MDAwWjCCAQcxMjAwBgoJkiaJk/IsZAEBDCJ1cm46bXJuOm1jbDpjYTptYXJpdGlt
+ZWNsb3VkLWlkcmVnMQswCQYDVQQGEwJESzEQMA4GA1UECAwHRGVubWFyazETMBEG
+A1UEBwwKQ29wZW5oYWdlbjEgMB4GA1UECgwXTWFyaXRpbWVDbG91ZCBVbml0IFRl
+c3QxIDAeBgNVBAsMF01hcml0aW1lQ2xvdWQgVW5pdCBUZXN0MTIwMAYDVQQDDClN
+YXJpdGltZUNsb3VkIFVuaXQgVGVzdCBJZGVudGl0eSBSZWdpc3RyeTElMCMGCSqG
+SIb3DQEJARYWaW5mb0BtYXJpdGltZWNsb3VkLm5ldDB2MBAGByqGSM49AgEGBSuB
+BAAiA2IABNfcAsujrQ6GUDnMjKy0Uo0RR00eRn+2vVzb6687+Ozk0WQSlaitKU6G
+3RlPKniWUU0JBsXCk1QY4oPPy6TmnY2z8tQmKra52dWVwPSYsrnvHyKnjOEnKMKu
+u4GXJAdasqOB0zCB0DAPBgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIB5jAf
+BgNVHSMEGDAWgBTys67EaTpHuOixcyLdCrz1vi5v3zAdBgNVHQ4EFgQUhYEdnD6m
+UQHPbIA3WrBepBvwJYMwbQYDVR0fBGYwZDBioGCgXoZcaHR0cHM6Ly9zdGFnaW5n
+LWFwaS5tYXJpdGltZWNsb3VkLm5ldC94NTA5L2FwaS9jZXJ0aWZpY2F0ZXMvY3Js
+L3Vybjptcm46bWNsOmNhOm1hcml0aW1lY2xvdWQwCgYIKoZIzj0EAwIDZwAwZAIw
+JZdNXFvE9HHo3lJqFk779HXXzRKlPXXtMMS5BP511c11y8c316kGohhEdmL4WLht
+AjAngKHJPREwbqgrikHsdjm75XopIkctYdKm4qLvuw+NoB22e9YWGRCwrVa8PQhl
+oZ8=
+-----END CERTIFICATE-----
