+++ mcp-pki/src/test/resources/ca/mc-root-cert.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDujCCAz+gAwIBAgIUDhlcvo7GkIJ5Z9aJX2f9PuSchgowCgYIKoZIzj0EAwIw
+gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
+bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
+CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
+dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
+Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM3WhcNMjQxMjMxMjMw
+MDAwWjCB0jELMAkGA1UEBhMCREsxEDAOBgNVBAgMB0Rlbm1hcmsxEzARBgNVBAcM
+CkNvcGVuaGFnZW4xIDAeBgNVBAoMF01hcml0aW1lQ2xvdWQgVW5pdCBUZXN0MSAw
+HgYDVQQLDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDExMC8GA1UEAwwoTWFyaXRp
+bWVDbG91ZCBVbml0IFRlc3QgUm9vdCBDZXJ0aWZpY2F0ZTElMCMGCSqGSIb3DQEJ
+ARYWaW5mb0BtYXJpdGltZWNsb3VkLm5ldDB2MBAGByqGSM49AgEGBSuBBAAiA2IA
+BIBZiJcroN0p+tx+KjQC47lyPIAul9jdj4mT5sFbskcIWlazhGTpC42Fctsikox8
+jIxjfGQCLIr15/VDymMLM8PT2gYXjgWAWCsLgvFHFvwSTlvpd4vbewY/nBJgOMz1
+haOB0zCB0DAPBgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIB5jAfBgNVHSME
+GDAWgBTys67EaTpHuOixcyLdCrz1vi5v3zAdBgNVHQ4EFgQU8rOuxGk6R7josXMi
+3Qq89b4ub98wbQYDVR0fBGYwZDBioGCgXoZcaHR0cHM6Ly9zdGFnaW5nLWFwaS5t
+YXJpdGltZWNsb3VkLm5ldC94NTA5L2FwaS9jZXJ0aWZpY2F0ZXMvY3JsL3Vybjpt
+cm46bWNsOmNhOm1hcml0aW1lY2xvdWQwCgYIKoZIzj0EAwIDaQAwZgIxAOWDOybk
+7mCQEcXae3DFfFhTZ9KtUg9WwoneRrcvyNYGrVSwolumWJbs6EpATKxyZQIxALs4
+WqZr+pBKHQpJuCvdJ/fDWSrHIb0kUS0B27LNNVbTODzE35mKGptcGFkbVIHwfA==
+-----END CERTIFICATE-----
