+++ mcp-pki/src/test/resources/ca/mc-truststore.jks
