+++ mcp-pki/src/test/resources/ca/root-ca-keystore.jks
