+++ mcp-pki/src/test/resources/mc-truststore-password-is-changeit.jks
