+++ mcp-pki/src/test/resources/thc-cert-nginx-format.pem
