+++ mcp-pki/src/test/resources/thc-private.pem
