+++ pom.xml
@@ -1,23 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
     <groupId>net.maritimeconnectivity.pki</groupId>
-    <artifactId>mcp-pki</artifactId>
+    <artifactId>mcp-pki-parent</artifactId>
+    <name>${project.artifactId}</name>
     <version>1.1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
-    <name>mcp-pki</name>
+    <packaging>pom</packaging>
     <url>https://maritimeconnectivity.net</url>
-    <description>PKI library for the MCP</description>
+    <description>Parent project for the MCP PKI library</description>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <skip.signing>true</skip.signing>
-        <bouncycastle.version>1.70</bouncycastle.version>
+        <maven.compiler.release>11</maven.compiler.release>
         <sl4j.version>1.7.33</sl4j.version>
         <lombok.version>1.18.22</lombok.version>
-        <maven.compiler.release>11</maven.compiler.release>
+        <skip.signing>true</skip.signing>
+        <verbose.delombok>false</verbose.delombok>
     </properties>
+
+    <modules>
+        <module>mcp-pki</module>
+        <module>mcp-pki-cli</module>
+    </modules>
 
     <licenses>
         <license>
@@ -49,47 +55,15 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.bouncycastle</groupId>
-            <artifactId>bcpkix-jdk15on</artifactId>
-            <version>${bouncycastle.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.bouncycastle</groupId>
-            <artifactId>bcprov-jdk15on</artifactId>
-            <version>${bouncycastle.version}</version>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>${sl4j.version}</version>
         </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
             <version>${lombok.version}</version>
             <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-            <version>${sl4j.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-simple</artifactId>
-            <version>${sl4j.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>commons-cli</groupId>
-            <artifactId>commons-cli</artifactId>
-            <version>1.5.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-engine</artifactId>
-            <version>5.8.2</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.junit.platform</groupId>
-            <artifactId>junit-platform-launcher</artifactId>
-            <version>1.8.2</version>
-            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -169,33 +143,9 @@
                             <goal>delombok</goal>
                         </goals>
                         <configuration>
-                            <verbose>true</verbose>
+                            <verbose>${verbose.delombok}</verbose>
                             <sourceDirectory>${project.basedir}/src/main/java</sourceDirectory>
                         </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <version>3.3.0</version>
-                <configuration>
-                    <descriptorRefs>
-                        <descriptorRef>jar-with-dependencies</descriptorRef>
-                    </descriptorRefs>
-                    <archive>
-                        <manifest>
-                            <mainClass>net.maritimeconnectivity.pki.Main</mainClass>
-                        </manifest>
-                    </archive>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>make-assembly</id> <!-- this is used for inheritance merges -->
-                        <phase>package</phase> <!-- bind to the packaging phase -->
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
                     </execution>
                 </executions>
             </plugin>
@@ -262,4 +212,5 @@
             </build>
         </profile>
     </profiles>
+
 </project>
