--- src/main/resources/META-INF/LICENSE
@@ -1,1 +1,0 @@
-../../../../LICENSE
