--- src/test/resources/ca/mc-iala-cert.pem
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDszCCAzigAwIBAgIUEPYUxwGeLgowuVUYdiX9n7SxR68wCgYIKoZIzj0EAwIw
-gdIxCzAJBgNVBAYTAkRLMRAwDgYDVQQIDAdEZW5tYXJrMRMwEQYDVQQHDApDb3Bl
-bmhhZ2VuMSAwHgYDVQQKDBdNYXJpdGltZUNsb3VkIFVuaXQgVGVzdDEgMB4GA1UE
-CwwXTWFyaXRpbWVDbG91ZCBVbml0IFRlc3QxMTAvBgNVBAMMKE1hcml0aW1lQ2xv
-dWQgVW5pdCBUZXN0IFJvb3QgQ2VydGlmaWNhdGUxJTAjBgkqhkiG9w0BCQEWFmlu
-Zm9AbWFyaXRpbWVjbG91ZC5uZXQwHhcNMTcwNDIxMDYxNzM5WhcNMjQxMjMxMjMw
-MDAwWjCByzEjMCEGCgmSJomT8ixkAQEME3Vybjptcm46bWNsOmNhOmlhbGExCzAJ
-BgNVBAYTAkZSMQ4wDAYDVQQIDAVQYXJpczEbMBkGA1UEBwwSU3QgR2VybWFpbiBl
-biBMYXllMRcwFQYDVQQKDA5JQUxBIFVuaXQgVGVzdDEXMBUGA1UECwwOSUFMQSBV
-bml0IFRlc3QxFTATBgNVBAMMDElBTEEgTWVtYmVyczEhMB8GCSqGSIb3DQEJARYS
-aW5mb0BpYWxhLWFpc20ub3JnMHYwEAYHKoZIzj0CAQYFK4EEACIDYgAEtzVUt1Yu
-EnK3wYGtKdVptHqhPHxVwI9/+qk+6HabBywrbSf9+3v8NC4AXIOFqraanCbgY46j
-W8qB7+UCpbE3BawRn5ShDlsUlcdX4UOLIHCNO1Nq/7Nj90wMlbSCOXgro4HTMIHQ
-MA8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgHmMB8GA1UdIwQYMBaAFPKz
-rsRpOke46LFzIt0KvPW+Lm/fMB0GA1UdDgQWBBSCytuLNaEyJFo4oYVnhk7xG7ON
-4jBtBgNVHR8EZjBkMGKgYKBehlxodHRwczovL3N0YWdpbmctYXBpLm1hcml0aW1l
-Y2xvdWQubmV0L3g1MDkvYXBpL2NlcnRpZmljYXRlcy9jcmwvdXJuOm1ybjptY2w6
-Y2E6bWFyaXRpbWVjbG91ZDAKBggqhkjOPQQDAgNpADBmAjEAvngPvblrjcSwp6nF
-mkFhDHdCQOCqaOqecqT/zUylDVmqSp/hkchE29RMI4cXwZQxAjEAuH52vQonHhl5
-sctrq85U3e9RmTDEau5TOgmmsliGx7JJglSycsE1m0mLQh1t4Tvr
------END CERTIFICATE-----
